ParanoidGlyph: Don't show charging animation on wiggle if screen is on
The animation had been claused to only be played when the screen is
facing down, but when lying down on a bed with the charger connected,
the charging animations continue to be triggered. The revised behavior
dictates that the wiggle animation will only be activated when the
screen is off or in a non-interactive state, such as Always On Display.

Change-Id: Id555276c8694dc88f5e55d32a454ce8fb11d90e8